# DEV: `@babel/plugin-proposal-decorators` -> `decorator-transforms`