--- a/app/assets/javascripts/dialog-holder/index.js
+++ b/app/assets/javascripts/dialog-holder/index.js
@@ -3,6 +3,23 @@
 module.exports = {
   name: require("./package").name,
 
+  options: {
+    babel: {
+      plugins: [
+        [
+          require.resolve("decorator-transforms"),
+          {
+            runEarly: true,
+          },
+        ],
+      ],
+    },
+
+    "ember-cli-babel": {
+      disableDecoratorTransforms: true,
+    },
+  },
+
   isDevelopingAddon() {
     return true;
   },