--- a/app/assets/javascripts/float-kit/index.js
+++ b/app/assets/javascripts/float-kit/index.js
@@ -2,7 +2,22 @@
 
 module.exports = {
   name: require("./package").name,
-  options: {},
+  options: {
+    babel: {
+      plugins: [
+        [
+          require.resolve("decorator-transforms"),
+          {
+            runEarly: true,
+          },
+        ],
+      ],
+    },
+
+    "ember-cli-babel": {
+      disableDecoratorTransforms: true,
+    },
+  },
   isDevelopingAddon() {
     return true;
   },