decorator-transforms (https://github.com/ef4/decorator-transforms) is a modern replacement for babel's plugin-proposal-decorators. It provides a decorator implementation using modern browser features, without needing to enable babel's full suite of class feature transformations.

Initial tests suggest:
- Our largest JS chunk drops from 4.38MB (991.71kb gzip) to 4.18MB (956.49kb gzip), so that's a 3.5% saving on gzip size

- Tachometer benchmarks suggest a 1-2% speedup in the time taken to boot the app and render the topic list

```
┌───────────┬─────────────────────┬─────────────────┬─────────────────┐
│ Benchmark │            Avg time │       vs before │        vs after │
├───────────┼─────────────────────┼─────────────────┼─────────────────┤
│ before    │ 220.22ms - 222.21ms │                 │          slower │
│           │                     │        -        │         1% - 2% │
│           │                     │                 │ 1.39ms - 4.09ms │
├───────────┼─────────────────────┼─────────────────┼─────────────────┤
│ after     │ 217.57ms - 219.39ms │          faster │                 │
│           │                     │         1% - 2% │        -        │
│           │                     │ 1.39ms - 4.09ms │                 │
└───────────┴─────────────────────┴─────────────────┴─────────────────┘
```

Currently installs decorator-transforms from a fork which includes [two pending fixes](https://github.com/ef4/decorator-transforms/pulls)

<!-- NOTE: All pull requests should have tests (rspec in Ruby, qunit in JavaScript). If your code does not include test coverage, please include an explanation of why it was omitted. -->