@Arkshine 2024-05-09
I just wanted to let you know that In a theme component, I was doing the following:

```js
api.modifyClass("service:chat-state-manager", {
  @tracked isChatSidebarActive,
  ...
});
```

However, after this PR, it no longer works, and `isChatSidebarActive` returns a function.
```js
ƒ () {
      return propertyDesc;
}
```

Maybe it was wrong from the start doing that, especially in a native class. (if https://github.com/discourse/discourse/pull/24957 is merged, I would assume it would be okay)
Using `isChatSidebarActive: tracked({ value: false }),` instead seems to work at the moment.